Exercises 4.18-4.22: token based authorization

+++ part4/bloglist/controllers/blogs.js
@@ -1,4 +1,5 @@
 const blogRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const mongoose = require('mongoose')
 const User = require('../models/user')
@@ -15,8 +16,13 @@ blogRouter.post('/', async (request, response) => {
     body.likes = 0
   }
 
-  const users = await User.find({})
-  const user = users[0]
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
     title: body.title,
@@ -34,8 +40,22 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  blog = await Blog.findById(request.params.id)
+
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({
+      error:
+        'Permission denied: Only the user who added the blog can delete it',
+    })
+  }
 })
 
 blogRouter.put('/:id', async (request, response) => {
